+++ middleware.go
@@ -7,7 +7,6 @@ import (
 	"github.com/twistingmercury/telemetry/v2/logging"
 	"github.com/twistingmercury/telemetry/v2/tracing"
 	"regexp"
-	"strconv"
 	"strings"
 	"time"
 
@@ -68,63 +67,82 @@ var (
 	nspace          string
 )
 
-// Initialize preps the middleware.
+//// Initialize preps the middleware.
+////
+//// Deprecated: use GetMetricsMiddleware, GetTracingMiddleware, and GetLoggingMiddleware which do not require separate initialization.
+//func Initialize(registry *prometheus.Registry, namespace, apiname string) error {
+//	switch {
+//	case registry == nil:
+//		return errors.New("registry is nil")
+//	case strings.TrimSpace(namespace) == "":
+//		return errors.New("namespace is empty")
+//	case strings.TrimSpace(apiname) == "":
+//		return errors.New("apiname is empty")
+//	}
 //
-// Deprecated: use GetMetricsMiddleware, GetTracingMiddleware, and GetLoggingMiddleware which do not require separate initialization.
-func Initialize(registry *prometheus.Registry, namespace, apiname string) error {
-	switch {
-	case registry == nil:
-		return errors.New("registry is nil")
-	case strings.TrimSpace(namespace) == "":
-		return errors.New("namespace is empty")
-	case strings.TrimSpace(apiname) == "":
-		return errors.New("apiname is empty")
-	}
-
-	reg = registry
-	nspace = namespace
-	apiName = apiname
-	concurrentCalls, totalCalls, callDuration = Metrics()
-	reg.MustRegister(concurrentCalls, totalCalls, callDuration)
-	return nil
-}
+//	reg = registry
+//	nspace = namespace
+//	apiName = apiname
+//	concurrentCalls, totalCalls, callDuration = Metrics()
+//	reg.MustRegister(concurrentCalls, totalCalls, callDuration)
+//	return nil
+//}
 
 // Telemetry returns middleware that will instrument and trace incoming requests.
 //
 // Deprecated: use GetMetricsMiddleware, GetTracingMiddleware, and GetLoggingMiddleware to get separated middleware for metrics, tracing, and logging.
-func Telemetry() gin.HandlerFunc {
+//func Telemetry() gin.HandlerFunc {
+//	return func(c *gin.Context) {
+//		path := c.Request.URL.Path
+//		method := c.Request.Method
+//		var elapsedTime float64
+//		var statusCode string
+//		concurrentCalls.WithLabelValues(path, method).Inc()
+//		defer func() {
+//			concurrentCalls.WithLabelValues(path, method).Dec()
+//			callDuration.WithLabelValues(path, method, statusCode).Observe(elapsedTime)
+//			totalCalls.WithLabelValues(path, method, statusCode).Inc()
+//		}()
+//
+//		spanName := fmt.Sprintf("%s: %s", c.Request.Method, c.Request.URL.Path)
+//		parentCtx := tracing.ExtractContext(c.Request.Context(), propagation.HeaderCarrier(c.Request.Header))
+//		childCtx, span := tracing.Start(parentCtx, spanName, oteltrace.SpanKindServer, semconv.HTTPRoute(spanName))
+//		c.Request = c.Request.WithContext(childCtx)
+//		defer span.End()
+//
+//		before := time.Now()
+//		c.Next()
+//		elapsedTime = float64(time.Since(before)) / float64(time.Millisecond)
+//
+//		logRequest(c, elapsedTime)
+//		code, desc := SpanStatus(c.Writer.Status())
+//		span.SetStatus(code, desc)
+//
+//		statusCode = strconv.Itoa(c.Writer.Status())
+//	}
+//}
+
+// Logging returns the logging middleware
+func Logging(excludePaths ...string) gin.HandlerFunc {
 	return func(c *gin.Context) {
 		path := c.Request.URL.Path
-		method := c.Request.Method
-		var elapsedTime float64
-		var statusCode string
-		concurrentCalls.WithLabelValues(path, method).Inc()
-		defer func() {
-			concurrentCalls.WithLabelValues(path, method).Dec()
-			callDuration.WithLabelValues(path, method, statusCode).Observe(elapsedTime)
-			totalCalls.WithLabelValues(path, method, statusCode).Inc()
-		}()
 
-		spanName := fmt.Sprintf("%s: %s", c.Request.Method, c.Request.URL.Path)
-		parentCtx := tracing.ExtractContext(c.Request.Context(), propagation.HeaderCarrier(c.Request.Header))
-		childCtx, span := tracing.Start(parentCtx, spanName, oteltrace.SpanKindServer, semconv.HTTPRoute(spanName))
-		c.Request = c.Request.WithContext(childCtx)
-		defer span.End()
+		if containsPath(excludePaths, path) {
+			c.Next()
+			return
+		}
+		var elapsedTime float64
 
 		before := time.Now()
 		c.Next()
 		elapsedTime = float64(time.Since(before)) / float64(time.Millisecond)
 
 		logRequest(c, elapsedTime)
-		code, desc := SpanStatus(c.Writer.Status())
-		span.SetStatus(code, desc)
-
-		statusCode = strconv.Itoa(c.Writer.Status())
 	}
 }
 
 // PrometheusMetrics returns the metrics middleware used by the Prometheus software.
-func PrometheusMetrics(registry *prometheus.Registry, namespace string, apiname string, options MetricsOptions) gin.HandlerFunc {
+func PrometheusMetrics(registry *prometheus.Registry, namespace string, apiname string, excludePaths ...string) gin.HandlerFunc {
 	switch {
 	case registry == nil:
 		panic("registry is nil")
@@ -143,8 +161,13 @@ func PrometheusMetrics(registry *prometheus.Registry, namespace string, apiname 
 
 	return func(c *gin.Context) {
 		path := c.Request.URL.Path
-		method := c.Request.Method
 
+		if containsPath(excludePaths, path) {
+			c.Next()
+			return
+		}
+
+		method := c.Request.Method
 		var elapsedTime float64
 		var statusCode string
 		concurrentCalls.WithLabelValues(path, method).Inc()
@@ -188,12 +211,11 @@ func Metrics() (*prometheus.GaugeVec, *prometheus.CounterVec, *prometheus.Histog
 }
 
 // OtelTracing returns the tracing middleware.
-func OtelTracing(options TracingOptions) gin.HandlerFunc {
+func OtelTracing(excludePaths ...string) gin.HandlerFunc {
 	return func(c *gin.Context) {
 		path := c.Request.URL.Path
 
-		_, pathFound := options.ExcludedPaths[strings.ToLower(path)]
-		if pathFound {
+		if containsPath(excludePaths, path) {
 			c.Next()
 			return
 		}
@@ -208,19 +230,6 @@ func OtelTracing(options TracingOptions) gin.HandlerFunc {
 
 		code, desc := SpanStatus(c.Writer.Status())
 		span.SetStatus(code, desc)
-	}
-}
-
-// Logging returns the logging middleware
-func Logging(options LoggingOptions) gin.HandlerFunc {
-	return func(c *gin.Context) {
-		var elapsedTime float64
-
-		before := time.Now()
-		c.Next()
-		elapsedTime = float64(time.Since(before)) / float64(time.Millisecond)
-
-		logRequest(c, elapsedTime)
 	}
 }
 
@@ -364,6 +373,18 @@ func ParseUserAgent(rawUserAgent string) (args map[string]any) {
 		args[UserAgentBrowserVersion] = ua.Version
 	}
 	return
+}
+
+func containsPath(paths []string, exclusion string) bool {
+	if len(paths) == 0 {
+		return false
+	}
+	for _, path := range paths {
+		if path == exclusion {
+			return true
+		}
+	}
+	return false
 }
 
 func normalize(name string) string {